@@ -19,6 +19,8 @@ public class Preferences {
     private static final String PREFS_HOS = "PREFS_HOS";
     private static final String PREFS_ACTIVE_DRIVERS = "PREFS_ACTIVE_DRIVERS";
     private static final String PREFS_NAVIGATION_STATE = "PREFS_NAVIGATION_STATE";
+    private static final String PREFS_IDENTITY_PROVIDER_SEND_JWT = "PREFS_IDENTITY_PROVIDER_SEND_JWT";
+    private static final String PREFS_IDENTITY_PROVIDER_TOKEN = "PREFS_IDENTITY_PROVIDER_TOKEN";
 
     public static SharedPreferences getPreferences(Context context) {
         return context.getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE);
@@ -75,7 +77,7 @@ public class Preferences {
     public static ArrayList<IdentityContract.Driver> getActiveDrivers(Context context) {
         String drivers = getString(context, PREFS_ACTIVE_DRIVERS, null);
         ArrayList<IdentityContract.Driver> activeDrivers = null;
-        if(drivers != null) {
+        if (drivers != null) {
             java.lang.reflect.Type listType = new TypeToken<ArrayList<IdentityContract.Driver>>() {
             }.getType();
             activeDrivers = new Gson().fromJson(drivers, listType);
@@ -91,5 +93,26 @@ public class Preferences {
 
     public static boolean getNavigationState(Context context) {
         return getPreferences(context).getBoolean(PREFS_NAVIGATION_STATE, false);
+    }
+
+    public static void setIdentityResponseJWT(Context context, boolean isJWT) {
+        SharedPreferences.Editor editor = getPreferencesEditor(context);
+        editor.putBoolean(PREFS_IDENTITY_PROVIDER_SEND_JWT, isJWT);
+        editor.commit();
+    }
+
+
+    public static boolean getIdentityResponseAsJWT(Context context) {
+        return getPreferences(context).getBoolean(PREFS_IDENTITY_PROVIDER_SEND_JWT, false);
+    }
+
+    public static void setIdentityResponseToken(Context context, String token) {
+        SharedPreferences.Editor editor = getPreferencesEditor(context);
+        editor.putString(PREFS_IDENTITY_PROVIDER_TOKEN, token);
+        editor.commit();
+    }
+
+    public static String getIdentityResponseToken(Context context) {
+        return getPreferences(context).getString(PREFS_IDENTITY_PROVIDER_TOKEN, "");
     }
 }
